requirements.txt
# https://github.com/Python-Markdown/markdown
markdown>=3.7, <4.0
bs4>=0.0.2, <0.0.3
pyyaml>=6.0.2, <7.0.0
jinja2>=3.1.4, <4.0.0